+++ app/src/main/java/io/github/regbl/migrainetracker/data/UserDao.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface UserDao {
     @Query("SELECT * FROM user_table WHERE id = :userId")
-    fun getUser(userId: Long): Flow<User>
+    fun getUser(userId: Int): Flow<User>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertUser(user: User)
